requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
psycopg[binary]>=3.2.10
python-dotenv==1.0.0
